[fix]작성하기수정하기  alert창띄우기 삭제 여부 묻기

+++ src/pages/community/[id].js
@@ -91,7 +91,10 @@ const Community = () => {
   const [isEditMode, setIsEditMode] = useState(false);
 
   const deletePostHandler = (postId) => {
-    deletePost(postId);
+    const confirmed = window.confirm('정말로 삭제하시기를 원하십니까?');
+    if (confirmed) {
+      deletePost(postId);
+    }
   };
 
   const [like, setLike] = useState(postLikeMine.like);
